Restricted access for creating, updating and deleting rooms with login_required

+++ base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -9,6 +10,9 @@ from .models import Room, Topic;
 from .forms import RoomForm
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect("home")
 
     if request.method == "POST":
         username = request.POST.get("username")
@@ -69,16 +73,30 @@ def createRoom(request):
 
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse("You shall not PASS!!")
+
+    if request.method == "POST":
+        form = RoomForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect("home")
 
     context = {"form": form}
 
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host :
+        return HttpResponse("You shall not PASS!!")
 
     if request.method == "POST":
         room.delete()
